Move dumpCheckpoint to universal ps
